refactor: further subdivide UserItem to extract info component
